+++ Kanbardoo.Application/TaskUseCases/DeleteTaskUseCase.cs
@@ -1,10 +1,15 @@
-﻿using Kanbardoo.Application.Constants;
+﻿using Kanbardoo.Application.Authorization.PolicyContracts;
+using Kanbardoo.Application.Constants;
 using Kanbardoo.Application.Contracts.TaskContracts;
 using Kanbardoo.Application.Results;
+using Kanbardoo.Domain.Authorization;
 using Kanbardoo.Domain.Entities;
 using Kanbardoo.Domain.Repositories;
 using Kanbardoo.Domain.Validators;
+using Microsoft.AspNetCore.Http;
 using System.Net;
+using System.Security.Claims;
+using System.Threading.Tasks;
 using ILogger = Serilog.ILogger;
 
 namespace Kanbardoo.Application.TaskUseCases;
@@ -14,14 +19,20 @@ public class DeleteTaskUseCase : IDeleteTaskUseCase
     private readonly ILogger _logger;
     private readonly IUnitOfWork _unitOfWork;
     private readonly KanTaskIdToDeleteValidator _kanTaskIdToDeleteValidator;
+    private readonly ITaskMembershipPolicy _taskMembershipPolicy;
+    private readonly int _userID;
 
     public DeleteTaskUseCase(ILogger logger,
                            IUnitOfWork unitOfWork,
-                           KanTaskIdToDeleteValidator kanTaskIdToDeleteValidator)
+                           KanTaskIdToDeleteValidator kanTaskIdToDeleteValidator,
+                           ITaskMembershipPolicy taskMembershipPolicy,
+                           IHttpContextAccessor contextAccessor)
     {
         _logger = logger;
         _unitOfWork = unitOfWork;
         _kanTaskIdToDeleteValidator = kanTaskIdToDeleteValidator;
+        _taskMembershipPolicy = taskMembershipPolicy;
+        _userID = int.Parse(contextAccessor.HttpContext!.User.FindFirstValue(KanClaimName.ID)!);
     }
 
     public async Task<Result> HandleAsync(int id)
@@ -33,9 +44,16 @@ public class DeleteTaskUseCase : IDeleteTaskUseCase
             return Result.ErrorResult(ErrorMessage.TaskWithIDNotExist);
         }
 
+        var authorizationResult = await _taskMembershipPolicy.Authorize(id);
+        if (!authorizationResult.IsSuccess)
+        {
+            return authorizationResult;
+        }
+
         try
         {
             await _unitOfWork.TaskRepository.DeleteAsync(id);
+            await _unitOfWork.UserTasksRepository.DeleteAsync(new KanUserTask { UserID = _userID, TaskID = id});
             await _unitOfWork.SaveChangesAsync();
 
             return Result.SuccessResult();
@@ -45,6 +63,5 @@ public class DeleteTaskUseCase : IDeleteTaskUseCase
             _logger.Error($"Internal server error: {id} \n\n {ex}");
             return Result.ErrorResult(ErrorMessage.InternalServerError, HttpStatusCode.InternalServerError);
         }
-        
     }
 }
